feat(FS-2): move from entity to domain and add port and service
