+++ internal/core/domain/category.go
@@ -1,4 +1,4 @@
-package entity
+package domain
 
 import (
   "time"
