+++ internal/core/domain/product_test.go
@@ -1,4 +1,4 @@
-package entity
+package domain
 
 import (
   "time"
